+++ src/SFA.DAS.FrameworkHelpers/RegexHelper.cs
@@ -11,6 +11,6 @@ public static partial class GeneratedRegexHelper
 
 public partial class ProjectNameRegexHelper
 {
-    [GeneratedRegex(@"SFA\.DAS\..*\.(Integration|PaymentProcess|API|UI|E2E)Tests")]
+    [GeneratedRegex(@"SFA\.DAS\..*\.(DataPreparation|Integration|PaymentProcess|API|UI|E2E)Tests")]
     public static partial Regex ProjectNameRegex();
 }
