+++ src/SFA.DAS.PayeCreation.DataPreparationTests/Project/ObjectContextExtension.cs
@@ -1,0 +1,12 @@
+﻿using SFA.DAS.FrameworkHelpers;
+
+namespace SFA.DAS.PayeCreation.DataPreparationTests.Project;
+
+public static class ObjectContextExtension
+{
+    #region Constants
+    private const string NoOfPayeRequested = "****NoOfPayeRequested****";
+    #endregion
+
+    public static void SetNoOfPayeRequested(this ObjectContext objectContext, int value) => objectContext.Set(NoOfPayeRequested, value);
+}
