+++ src/SFA.DAS.PayeCreation.DataPreparationTests/Project/PayeCreationConfig.cs
@@ -1,0 +1,14 @@
+﻿namespace SFA.DAS.PayeCreation.DataPreparationTests.Project;
+
+public class PayeCreationConfig
+{
+    public string Duration { get; set; }
+
+    public string LevyPerMonth { get; set; }
+
+    public string NoOfLevy { get; set; }
+
+    public string NoOfNonLevy { get; set; }
+
+    public string NoOfAornNonLevy { get; set; }
+}
