+++ src/SFA.DAS.PayeCreation.DataPreparationTests/Project/ScenarioContextExtension.cs
@@ -1,0 +1,14 @@
+﻿using TechTalk.SpecFlow;
+
+namespace SFA.DAS.PayeCreation.DataPreparationTests.Project;
+
+internal static class ScenarioContextExtension
+{
+    #region Constants
+    private const string PayeCreationConfigKey = "payecreationconfigkey";
+    #endregion
+
+    internal static void SetPayeCreationConfig(this ScenarioContext context, PayeCreationConfig value) => context.Set(value, PayeCreationConfigKey);
+
+    internal static PayeCreationConfig GetPayeCreationConfig(this ScenarioContext context) => context.Get<PayeCreationConfig>(PayeCreationConfigKey);
+}
