+++ src/SFA.DAS.PayeCreation.DataPreparationTests/SFA.DAS.PayeCreation.DataPreparationTests.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
-    <RootNamespace>SFA.DAS.PayeCreation.Tests</RootNamespace>
+    <RootNamespace>SFA.DAS.PayeCreation.DataPreparationTests</RootNamespace>
 	  <ErrorOnDuplicatePublishOutputFiles>false</ErrorOnDuplicatePublishOutputFiles>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
