--- src/SFA.DAS.PayeCreation.Tests/Project/PayeCreationConfig.cs
@@ -1,15 +1,0 @@
-﻿namespace SFA.DAS.PayeCreation.Project
-{
-    public class PayeCreationConfig
-    {
-        public string Duration { get; set; }
-
-        public string LevyPerMonth { get; set; }
-
-        public string NoOfLevy { get; set; }
-
-        public string NoOfNonLevy { get; set; }
-
-        public string NoOfAornNonLevy { get; set; }
-    }
-}
